Reorganised code into multiple files
